P_04/hyper/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudml-hypertune==0.1.0.dev6
flatbuffers==23.5.26
gast==0.5.4
google-api-core==2.12.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.33.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
keras==2.14.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
typing-extensions==4.8.0
urllib3==2.0.5
werkzeug==2.3.7
wrapt==1.14.1
zipp==3.17.0
